// src/src/app/services/models/ListCalification.ts
export class ListCalification {
    PoliticalGroupCode: string;
    PoliticalGroupName: string;
    PoliticalGroupImageUrl: string;
    CandidatesCount: number;
    OneOneValue: number;
    OneOnePoints: number;
    OneTwoValue: number;
    OneTwoPoints: number;
    OneThreeValue: number;
    OneThreePoints: number;
    OneFourValue: number;
    OneFourPoints: number;
    OneFiveValue: number;
    OneFivePoints: number;
    TwoOneValue: number;
    TwoOnePoints: number;
    TwoTwoValue: number;
    TwoTwoPoints: number;
    TwoThreeValue: number;
    TwoThreePoints: number;
    ThreeOneValue: number;
    ThreeOnePoints: number;
    ThreeTwoValue: number;
    ThreeTwoPoints: number;
    ThreeThreeValue: number;
    ThreeThreePoints: number;
    ThreeFourValue: number;
    ThreeFourPoints: number;

    ComponentOneAvgPoints: number;
    ComponentTwoAvgPoints: number;
    ComponentThreeAvgPoints: number;
    TotalPoints: number;
}